drf-react/backend/requirements.txt
Django==1.11.2
django-filter==1.0.4
djangorestframework==3.6.4
Markdown==2.6.9
psycopg2==2.6.1
gunicorn
